backend/requirements.txt
fastapi
uvicorn
google-analytics-data
supabase
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv
